Proof of concept for combining Transformer/Indexing Producers
They both do the same thing and the code was all the same except two
methods. This is one way I was able to combine them.
