+++ test/dpul_collections/indexing_pipeline/integration/figgy/indexing_integration_test.exs
@@ -15,7 +15,7 @@ defmodule DpulCollections.IndexingPipeline.Figgy.IndexingIntegrationTest do
 
     :telemetry.attach(
       "ack-handler-#{pid |> :erlang.pid_to_list()}",
-      [:indexing_producer, :ack, :done],
+      [:database_producer, :ack, :done],
       fn _event, _, _, _ -> send(pid, {:ack_done}) end,
       nil
     )
