+++ test/support/mock_figgy_indexing_producer.ex
@@ -16,7 +16,9 @@ defmodule MockFiggyIndexingProducer do
   @type state :: %{consumer_pid: pid(), test_runner_pid: pid(), indexing_producer_pid: pid()}
   @spec init({pid(), Integer}) :: {:producer, state()}
   def init({test_runner_pid, cache_version}) do
-    {:ok, indexing_producer_pid} = Figgy.IndexingProducer.start_link(cache_version)
+    {:ok, indexing_producer_pid} =
+      Figgy.DatabaseProducer.start_link({Figgy.IndexingProducer, cache_version})
+
     {:ok, consumer_pid} = MockConsumer.start_link(indexing_producer_pid)
 
     {:producer,
